refactoring, extract pino-logger and winston-logger to their own files
